profiles changed, structure of the project corrected
