+++ src/test/java/com/epam/tc/hw1/DivisionOperationTest.java
@@ -1,12 +1,14 @@
+package com.epam.tc.hw1;
+
 import static org.testng.AssertJUnit.assertEquals;
 
 import com.epam.tat.module4.Calculator;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
-public class SumOperationTest {
+public class DivisionOperationTest {
 
-    @DataProvider(name = "SumOperationTestDataProvider")
+    @DataProvider(name = "DivisionOperationTestDataProvider")
     public Object[][] testNumbers() {
         return new Object[][] {
             {5L, -5L},
@@ -15,12 +17,12 @@ public class SumOperationTest {
         };
     }
 
-    @Test(dataProvider = "SumOperationTestDataProvider")
-    public void shouldSumTwoLongNumbers(Long number1, Long number2) {
+    @Test(dataProvider = "DivisionOperationTestDataProvider")
+    public void shouldDivideTwoLongNumbers(Long number1, Long number2) {
         Calculator calculator = new Calculator();
 
-        Long actual = calculator.sum(number1, number2);
-        Long expected = number1 + number2;
+        Long actual = calculator.div(number1, number2);
+        Long expected = number1 / number2;
 
         assertEquals(expected, actual);
     }
