+++ src/test/resources/testng/defaultTests.xml
@@ -4,11 +4,12 @@
   <test name="defaultOperationsTest">
     <groups>
       <run>
-        <include name="default"/>
+        <include name="defaultOperationsTest"/>
       </run>
     </groups>
     <packages>
-      <package name="src/test/java"/>
+      <package name="com.epam.tc.hw1">
+      </package>
     </packages>
   </test>
 </suite>
